C#: Add file instead of generated location for extraction errors when possible
